Trying to teleport known constants into the argument positions of primitive operations (so they can then be removed from function argument lists) reveals a bug in recursion.  It seems that analyzing LETREC-through-Y-combinator leads to the analysis finding the least-commitment solution to the type equations, which is that every recursive function (that doesn't get fully unrolled because the structure it recurs over is known) returns the abstract-all value.  I am surprised that it took this long to get bitten by this bug; but I think it worked so far because the primitive procedures effectively do one step of type inference on their arguments, because the code generation does not actually need to know the shape of the parameters to primitives (it can emit code for them anyway).  Which means that the system will generate correct code as long as there are only primitive operations looking at the (nested) outputs of recursive functions.
I added an example with a compound procedure looking at the output of
a recursive function, and added code to the code generator to be picky
about the shapes of the arguments to primitive procedures.  Now all my
numeric recursion loops error out.

I think now I know why Siskind and Pearlmutter made LETREC be a
primitive special form in VLAD.  The question is, how should this be
fixed?