Starting to think about DVL, the Dys-Vunctional Language.
This would be a modification of VL to allow limited state; namely a
gensym procedure that does what you would expect and can be analysed
effectively.  The reason I want gensym is that it will allow me to
implement forward-mode AD at the client level without further problems
(and in particular without having to worry about confusing my
perturbations).  While I'm at it, it seems I can get the analysis to
understand i/o for free, but convincing the code generator to play
along properly may be somewhat difficult.