Oof, what a bug!  The value of abstract-hash changed on every garbage collection, because there was an eqv-hash sitting inside.
That was not such a big deal in itself, because the memoize cache
would standardize a particular value, and as long as all instances
if eqv-hash inside abstract values were protected by the memoize cache,
there would not be any missing hash table entries.

But then, greater fool I, I switched memoize to using
hash-table/intern!.  This was a mistake, because as I was to learn,
hash-table/intern! requires that the procedure used to compute the
intern value must not modify the table into which the thing is being
interned.  The punishment for violating this condition is misplaced
hash table entries, and occasional exceptions during interning,
which I promptly started to get.

The misplaced entries exposed the problem with abstract-hash, because
now the memoize cache did not reliably prevent the abstract-hash
function from being recomputed; but by saving some of the old values,
had the effect that sometimes values of abstract-hash generated at
different times, with GC in between, would be compared.  This caused
my analysis data structure to mysteriously drop bindings, leading
to exceptions in the code generator.

This has now been fixed, in a belt-and-suspenders kind of way: I
reverted the memoize cache to using hash-table/lookup, and also
rewrote abstract-hash to be invariant under GC.