Added some comments in the DVL source.
I am probably not the right person to write the DVL's documentation
anymore because the more I look at the code, the more it makes sense
to me, and the more I get the impression that the necessary comments
are already there, but they should not be read linearly in order to
understand what's going on.  Sometimes, in order to understand what a
procedure does, it is helpful to see where it is used and to read the
comments surrounding its call site.