Draft compilation of FOL to Stalin (namely, r4rs).
This will obviously need to be chewed on some more, but it at least
compiles and runs, on at least one example.  I am surprised that it
gets a drastically different answer from MIT Scheme; but that answer
was grossly wrong to begin with, so I don't know where to start
chasing the difference.