Draft SRA support for emitting closures to MIT Scheme.
Under current assumptions (namely that all escaping functions only
accept real numbers), the only interesting thing that needs to be done
to a lambda expression is that it needs to reconstruct the shape being
returned to Scheme.  A sub-lambda-expression is already the right
shape for Scheme, and needs no additional reconstruction.