Syntactic changes.
Don't intermix 'double' and 'real', use the latter throughout.
Introducing Real as a type synonym for Double requires hiding
Prelude.Real, hence the use of the NoImplicitPrelude language
extension).

Also, types and constructors live in different namespaces, so
that it is possible to have Bool constructor of the Expr data
type.