New implementation of the alpha-renamer.
Moved it to the file AlphaRn.hs.  The new approach turns Rename into a
monad transformer RenameT, which is applied to the monad Unique, so
that it is easier to get the state the renamer is carrying around (the
list of names known to be bound somewhere).  Also, we don't have to
(and it is wrong to) store actions of the type Unique Name in
expressions, so I am going to revert back to the simple representation
of expression that is not parametrized over the type of names.  We
also don't need Functor, Foldable, and Traversable instances (and they
are probably the wrong, or at least the useless, ones anyway).