Approximate A-normal form.
There is at least one reason why this code sucks: approximate A-normal
form has some invariants that could be captured in types, but we
ignore this and use the same Expr type.  It would be more appropriate,
but arguably even less elegant, to introduce a new type, AnfExpr,
corresponding to the grammar of expressions in approximate a-normal
form.  Doing so would require writing tedious conversions from AnfExpr
to Expr (unless we change all subsequent stages of the optimizer to
accept AnfExpr instead of Expr).