Execution functions get a place of honor in the stage definition language.
Execution functions are pretty special, because the clause has to be
before any execution function modifier clauses and because I want to
eta expand in order for the stage structure to pick up any REPL-zapped
changes to the execution function.  So it's reasonable to give them
their own place in the syntax for defining stages.