Explanations reveal more bugs.
The application case should be more careful not to replace an
application with an out-of-scope variable.  The way an out-of-scope
variable can arise as a candidate involves simplification of the
symbolic expression, as happens in the test case.  While we're at it,
we should simplify applications even when they are not replaced.