Stages machinery version 3.
It now appears cleaner to separate preparation from action.  The
preparation is when required properties may be created, etc, and the
action is just the raw pass.  Stage effects like generates, preserves,
and destroys modify the action step, because these things are
consequences of the actual pass.  The requires property, however,
modifies the preparation step, because massaging the input is
preparation.  Now the wrapping loop can just affect the action steps.