Abandon the old world order wrt checking invariants with unit tests.
I have recovered what testing goodness I could from the stage system.
The other invariants that the old world order used to test depend on
trying alternate stage orderings and verifying that they have the same
results as the ones actually tried.  I do not know how to do this in a
setting where the stage order is flexible and not hardcoded into the
unit test tools.