Hack the type-checker to collect types of subexpressions.
This is ugly: an eq-hash-table is supplied as an optional parameter,
and the main loop of CHECK-EXPRESSION-TYPES procedure is advised to
stick types it computes into that hash-table (if it's available).

Unfortunately, this doesn't always work: some expressions are symbols
and compare as equal even though their source positions are different.
Argh!