Augment the compiler to decorate expressions with their types using `the' form.
Unfortunately, this doesn't always work because the collected types
are not always correct: expressions that are symbols and occur in
different places in the source code will have the same type.

Should I write this compiler in Haskell?  My type-checker at least
doesn't have this problem (and doesn't require hacking).