Parsimony of (the ...) expressions.
Declaring every formal parameter (including those of exported
functions) and the return type of every procedure (at the call site)
suffices for SBCL to infer all the types that were explicitly declared
before, at least on the celestial mechanics example.  (Do we have a
test suite for this sort of thing?)  The advantage is that the
resulting .lisp file is that much smaller, and simpler to think about.
(Coincidentally, compilation with SBCL also runs faster.)