Some little helpers for chasing bugs in DVL.
These were inspired by a bug in the forthcoming draft of multilayer
perceptron inference, where I wrote (sum (dot v1 v2)) instead of
(dot v1 v2).  That took way too long to chase down, because:
1) The manifestation of the bug was a call to (error).
2) This (error) turned out to be inside pointwise2
   (case statements macro expand most annoyingly to the eye).
3) The pointwise2 was inside v:+, and the error was because it was
   being asked to add a gensym to a real.
4) The reason for that was that somebody tried to sum a bundle.
   (And bundles being lists, sum was happy to try).
5) The reason for that was that dot produced a bundle because I was
   differentiating the vectors whose dot product was wanted.

Another thing that was annoying was that the binding stack took me
through the guts of the macroexpansion of letrec.