Remove a slow redundant invocation of dead code elimination.
Perhaps I can put it back after eliminate-intraprocedural-dead-code
stops being quadratic (or is it cubic?) in the number of arguments
that get passed to a procedure.

This removal has no effect on the DVL test suite, but I expect it to
make an appreciable difference for speed of optimizing larger
examples.