Compute a map for the liveness of inputs as well as outputs by direct recursion instead of going through the dependency map.
The goal is to eliminate the troublesome overprecision that leaves
tombstones that do not get flushed.  To do this, treat all procedures
as strict in all arguments, instead of using the dependency map to
decide which arguments are needed at this call site and which ones are
not.  This commit computes the input liveness map by that rule; the
next will use that input liveness map when rewriting procedure
definitions.