Mitigate allocation inside estimate-space-usage.
The purpose is to accelerate the measurement process, and reduce the
impact of observation on that which is observed.  It is not possible
to count memory usage accurately without a keeping a record of visited
objects, which entails allocation, but it is possible to avoid all
other allocation, which I think I have now done.  The effect is that
estimating the space usage of a 52MW analysis object is now about five
times faster.

The method is basically manual loop fusion, to avoid allocating
intermediate lists, and source diving to figure out how to traverse
the fields of a record without allocating either the field name list
or the accessor procedures.