Simple-minded FOL-to-JavaScript compiler.
Code generation proceeds as if JavaScript were Scheme.

Because JavaScript lacks syntax for introduction of lexical bindings,
but has lambdas, LET and LET-VALUES are expressed using lambdas.
There are no multiple value returns, so these are simulated by
returning arrays.  As a consequence, the generated code will cons.
Also, the arithmetic is not unboxed.  The only benefit of running FOL
in this case are probably some CSE and dead code elimination.