Converting structures to vectors is now officially out of the pipeline.
It had to wait until this point because, even though it affects no
code if invoked after SRA examples that have no recursive types, it
did have the salutary effect of flushing the dead type declarations
that the rest of the pipeline left around.  This task is now taken
over by dead type elimination.