A unit test for a bug I have now -- SRA loops on recursive types.
Right now there is no way to construct a value of recursive type, so
this bug is just a discrepancy between the laxity of syntax the type
checker is willing to accept vs the generality in which SRA operates.
When I introduce types for escaping functions, however, these kinds of
things may become larger issues.